+++ src/view-engine/src/Factory.php
@@ -12,10 +12,10 @@ declare(strict_types=1);
 namespace Hyperf\ViewEngine;
 
 use Closure;
+use Hyperf\Macroable\Macroable;
 use Hyperf\Utils\Arr;
 use Hyperf\Utils\Contracts\Arrayable;
 use Hyperf\Utils\Str;
-use Hyperf\Utils\Traits\Macroable;
 use Hyperf\ViewEngine\Contract\EngineInterface;
 use Hyperf\ViewEngine\Contract\EngineResolverInterface;
 use Hyperf\ViewEngine\Contract\FactoryInterface;